Establishing inital framework
The entry point of the library is the Connection Interface. This
Interface exposes a number of operations that can be performed on
an SSH server.
Created a single implementation of the Connection Interface that uses
SFTP to interact with an SSH server.
Implementations of the Connection Interface are instantiated using the
ConnectionFactory.

Most methods in the Connection Interface call into methods of the same
name in the RemoteSession Interface. This is an internal Interface used
to interact with an SSH server.
Created an implementation of RemoteSession that uses JSch to communicate
with an SSH server.

Authentication with SSH servers is handled by the Authentication
Interface. Implementations of this Interface are responsible for
configuring authentication for a RemoteSession to use with an SSH server.
Currently only password authentication is supported, but key based is
coming.